fix:refactor handle save function
